Breaking Change Infrastructure
The implementation relies on a boolean `breaking` field on the dependencies struct and locked fields
on chart and terraform installation schemas.  When a version is pushed marked breaking, the version is installed but
not passed to its upgrade queue, and marked locked.  Any downstream dependencies of that version are blocked from
being installed, and thus will be sent to a deferred queue.

We'll simultaneously deliver a user to affected users giving them instructions on how to apply the change, ultimately
resolvign to a cli command which will unlock the installations and restart upgrade queues.  We also add templatable sections
to allow maintainers to give additional useful instructions for upgrades (maybe kubectl commands and so forth).