Add OIDC Login Auditability
We've already added the oidc_login table for us to track installation health, and I recently realized it could also be used for teams to monitor login activity for their app footprint as well.  Adding a gql query and an audit table for www to do this.  We can also add things like ip linking and such as in all our audits too.