Backpressured docker image scanning
We currently continuously scan images by running a cron that scrapes them and shoves them into rabbit.  This can cause pretty big load spikes on our workers, and it'd be ideal to have more control on the rate of work.  Using a genstage based flow will do that, by allowing us to control overall demand in the system and configure poll intervals better.

This still shoves them into rabbit, which could be removed as well in theory.