Add scan_completed_at for better docker scan bookkeeping
At the moment we have a table poller that marks scanned_at when it dequeues an image, but it can fail during scan and has to wait days to requeue.  Adding another timestamp to mark if it's been dequeued and scanned should allow us to immediately requeue it in the event of a failed scan.