Platform Payments Usage Accounting
The basic architecture works as follows:

* pubsub system automatically updates user/cluster counts on relevant events, and also marks the account as having been updated
* a cron scrapes accounts w/ that boolean set and syncs the current count to stripe on a daily basis.

This is meant primarily to minimize the number of invoices stripe creates and also solves for the fact that stripe's api is slow enough it's dangerous to run in high-volume transactions.

One additional note, I'm updating cluster delete transactionally, since it's done in a batch job there's a chance event draining is too slow and will miss a few.